moved the interactive page over to the explore the data page
